requirements.txt
altgraph==0.17.4
bottle==0.12.25
cffi==1.16.0
clr-loader==0.2.6
lxml==4.9.3
packaging==23.2
pefile==2023.2.7
proxy-tools==0.1.0
pycparser==2.21
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pythonnet==3.0.3
pywebview==4.4.1
pywin32-ctypes==0.2.2
typing_extensions==4.8.0
